github_copyfile/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
boto3
python-dotenv